--- a/benchmarks/CSharpTaskBenchmarks/CSharpTaskBenchmarks.csproj
+++ b/benchmarks/CSharpTaskBenchmarks/CSharpTaskBenchmarks.csproj
@@ -1,11 +1,11 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
-  <PropertyGroup>
-    <TargetFramework>net7.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-    <IsPackable>false</IsPackable>
-  </PropertyGroup>
+    <PropertyGroup>
+        <TargetFrameworks>net6.0;net7.0</TargetFrameworks>
+        <ImplicitUsings>enable</ImplicitUsings>
+        <Nullable>enable</Nullable>
+        <IsPackable>false</IsPackable>
+    </PropertyGroup>
 
-  <Import Project="..\..\.paket\Paket.Restore.targets" />
+    <Import Project="..\..\.paket\Paket.Restore.targets" />
 </Project>