--- a/benchmarks/FSharpBenchmarks/asyncCompletion.ps1
+++ b/benchmarks/FSharpBenchmarks/asyncCompletion.ps1
@@ -1,0 +1,1 @@
+dotnet run -c Release -f net7.0 -- --filter "*IcedTasks.Benchmarks.AsyncCompletionBenchmarks*"